Finish splitting rfds file
